Fix normalization of slashes in filenames
All paths were canonicalized internally, which includes filtering
sequences of slashes and converting backslashes to slashes.
Furthermore, when unpacking files, filenames are sanity checked
and rejected if they contain forward OR backward slashes.

This is a problem on Unix-like systems, where files containing
backslashes are a legitimate use case (*cough* SystemD *cough*).

This patch removes the backslash conversion from the canonicalization
and modifies the sanity check to reject backslashes only on Windows.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>