libsqfs: Add a sqfs_block_processor_create_ex function
This function creates a block processor from a structure describing it.
A stub implementation for the old sqfs_block_processor_create is added
that simply sets up such a struct and forwards the call.

The current version of the description struct only contains the exact
same parameters and a size field at the beginning.

This approach is supposed to make extending the range of parameters
easier without breaking ABI compatibillity.

Currently already planned are:
 - Adding a sqfs_file_t pointer to double-check when deduplicating
   fragments.
 - When the scanning code reaches a usable state, add the abillity
   to pass scanned fragment data, so the block processor can be used
   for appending to an existing image.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>