Fix: Move fragment consolidation back to block processor serial part
Keeping a list of fragments stored away in the current fragment block
and consolidating them in the thread pool takes them out of circulation.

If we have a lot of tiny fragments, this can lead to a situation where
all the limit is reached, but we cannot do anything, because we are
waiting for a block to complete, but they are all attached to the
current fragment block and the queue is empty.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>