Bump version numbers & changelog after 1.0.4 release
This commit imports the changes to the CHANGELOG.md in the backports
branch after the 1.0.4 release and updates the package scripts to
point to the new release.

In addition, the version numbers in configure.ac are update to more
closeley reflect reality, i.e. set the current so version and set the
package version to 1.1.0 which will be the next release from master.

The unreleased changes in master are also added to the CHANGELOG.md.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>