fstree_from_file: Implement basic file globbing
The base path is passed to the fstree_from_file function and in turn
to the individual callbacks.

The line parsing function is modified to allow '*' as mode, uid and gid
for specifically marked callbacks.

A glob callback is added that internally uses the fstree_from_dir scanning
functions in combination with a filter callback.

Directory scanning flags are parsed from the extra arguments before
interpreting it as a path fragment.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>