Fix: libfstree: add an assert the canonicalize_name return value
Since the canonicalize_name function only fails if the path
contains ".." and the one we are constructing from the scanned
fstree (built using canonicalized names), it should NEVER fail.

However, coverity does get concerned, because we are checking the
return value elesewhere. So do what we do at other, similar locations
and add an assert().

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>