Remove lz4 & zstd from corpus test
Relying on the output of a compressor to exactely match an expected
output is already not really a great idea, but for gzip, xz and lzo
it has worked remarkably well so far. Perhaps because those are rather
old and don't have much active development going on besides bug fixing.

On the other hand, lz4 and zstd which are much younger seem to have
more development going on and keep breaking between versions.

This commit removes the zstd & lz4 corpus tests.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>