Optionally use a pool allocator for rb-tree nodes
This commit restructures the rbtree code to optionally use a pool
allocator for the nodes. The option is made depenend on the presence
of a pre-processor flag.

To the configure script is added an option to enable/disable the use
of custom allocators. It makes sense to still allow the malloc/free
based routes for better ASAN based instrumentation.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>