Fix: allow concatenated xz streams
Some xz compressed tarballs (e.g. from kernel.org) are not made up of
a single xz stream, but rather contain several, independendly
compressed streams. In that case, the xz decompressor hits
an LZMA_STREAM_END early on and reports EOF. If you are lucky, the tar
reader bails (premature end-of-file). If you are unlucky, it happens
exactely between two records and is interpeted as regular end-of-file.

As this seems to be a normal use case for xz, it has a flag to just
read across the seams and only report end-of-stream if the action
is set to finish.

This commit adds the flag to the initialization propperly sets the
lzma_action depending on whether the underlying stream hit EOF or not.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>