Cleanup: table driven pax header parsing
Instead of having a long if-else-if chain, replace the PAX header
field parsing with a table driven approach.

Altough it is more code, it is hopefully more readable, maintainable,
extensible and it dedupliates some of the value parsing code.

The GNU.sparse parsing is left as is, because it requires
maintaining state.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>