Fix half done initialization of sqfs_dir_reader_open_dir
The sqfs_dir_reader_open_dir function tried to take a short-cut by
returning early if the target directory is empty. However, this left
some field unchanged from the previous directory.

If iterating over a directory and then deciding to enter a sub-directory
that happens to be empty, the directory reader will keep the settings
for the current directory. After calling sqfs_dir_reader_rewind, the
sub-directory will suddenly report the contents of the parent.

A similar check is added to the rewind function to not track back on
the meta data reader in that case.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>