Cleanup: remove struct stat from libtar
The idea was originally to use struct stat in the libfstree code, so
we can simply hose data read from a directory into the fstree_t. The
struct was then also used with libtar, for simpler interoperation,
but it turned out to introduce a lot of platform quirks and causes
more trouble than it's worth.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>