Add sort-file implementation
A `flags` field and `priority` are added to all file information
structs. A news fstree function is introduced for parsing a "sort-file".

Each line in the file is space separated, and has the following format:

  priority [flags] filename

Priority is a 64 bit number, flags are optional and filename can be put
in quotes if it is supposed to start or end with spaces. Single line
comments can be used.

The flags can be used to set block-processor flags (e.g. don't fragment,
or don't compress), as well as instructing the parser to use file
globbing to match the filename.

After parsing the file, the list of file info structure is sorted
according to the priority (default is 0) using a stable sort algorithm.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>