Add a wrapper for the main function on Windows
A macro and forward declaration are added to compat.h that rename
the main() function programs using compat.h into sqfs_tools_main.

An actual main() function is added to libcompat.a, that uses the
shell API to get the UTF-16 command line arguments, convert them
to UTF-8 and call sqfs_tools_main.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>