Fix: Directory insertion order on Windows
The fstree sorting code got removed recently, in favour of inserting
at the correct position. The Windows directory scanning code still
used it's own list insertion code instead of mknode (so it could
allocate and translate the directory entry name in-place), which
broke the sorting order.

This issue is fixed in this commit.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>