Fix: libfstream: don't fail on Windows when reading from a pipe
When piping the output of another program into tar2sqfs.exe, and
the source program terminates, tar2sqfs.exe gets an ERROR_BROKEN_PIPE
when the end is reached and it trys to pre-cache more data. This
commit adds a work around, to propperly handle this as and end-of-file
condition.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>