Remove builtin copy of zlib
On GNU/Linux, *BSD or MacOS we can simply use the system default
library. The copy was primarily only there for the Windows build.

The build script for Windows has now been adapted to download and
compile a shared library from a tarball.

This removes a huge chunk of code from the git tree as well as
the release tarballs. Additionally it gets rid of iffy things like
removing the Zlib copyright/version strings, so the libsquashfs DLL
doesn't export it.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>