Only use available CPUs
Not all CPUs may be available for the current process. Some CPUs
may be offline, others may not be included in the process affinity
mask. In such cases too many threads will be created, which will
then compete unnecessarily for CPU time.

Use sched_getaffinity() to determine the correct number of threads
to create.