filemap xattr: use file istream from libio instead of stdio FILE
The line-by-line reading function in the io library was specifically
added for this kind of use case, so we don't have to handle trimming
of lines, don't have to touch getline() and it's convoluted error
handling (return value -1 could mean either EOF or error).

The code that searches for '\n' and replaces it with '\0' can be
removed as well and a memory leak involving "line" is removed along
the way.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>