libio: Add an xfrm stream wrapper test
This basically re-uses the libxfrm pack/unpack tests, but runs
the data through a stream wrapper.

When de-compressing, we have a ridiculously tiny input buffer,
to force the wrapper to snort up the data in several attempts
until the it can decompress something. We read the data byte-by-byte
to force the wraper to internally cache the uncompressed data and
spoon feed it to us. It has to be completely transparent to us that
it internaly decompresses and also reads transparently across
concatenated streams.

When compressing, we only require that the output is smaller than
the input and not equal to it. We also require the wrapper to flush
the wrapped stream when it is flushed. We then test if the compressed
data can be unpacked again.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>