libsquashfs: remove the default block writer alignment feature
The idea of the block align feature was to allow micro-managing that
some files are forcefully aligned to 1k/4k ("device block") boundaries,
hoping to improve access time at the cost of data density. The feature
was not exposed in the tools for a long time and eventuall added to the
sort file. Measurement and experimentation showed, that it in fact
worsened the read performance on a test system with an old micro SD
card as the bottle neck.

The feature is removed, and if needed, can be brought back simply by
wrapping/sub-classing the default block writer, if need be..

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>