gensquashfs: Remove flag handling from scan_dir
The logic was previously there to allow recursing into directories
that exist in the fstree and the scanned directory, but not create
new directories. Because the dir_tree_iterator_t supports reporting
contents but not directories (and does DFS), we can simply check if
the parent of the directory exists. Under normal DFS, it must exist,
because we created it earlier. If we are skipping the directory
entries, it only exists if it was there earlier and we can skip
entries that have a non-existant parent, retainint the old behavior.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>