gensquashfs: use prefix functionality for glob_files target node
Piece together the prefix path and pass it to the iterator. That way,
we get the full target paths back from the iterator and can use those
directly in the callback for filtering.

We also no longer need the root node for fstree_from_dir (always tree
root) and the callback can no longer return an error state.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>