libio: istream_get_line: fix trailing CR sanitation
The istream_get_line function is supposed to remove the line break,
inclduing CR+LF if used. The previous implementation simply checked
the buffer when the LF was found, missing the case where the CR was
alraedy added and the LF is at the beginning of the newly loaded
data.

This commit modifies the check to inspect the line itself when it
was found, and remove the CR after the fact.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>