libtar: overhaul buffer management in istream_t wrapper
As soon as we no longer have any data to read, unlock/drop the
parent iterator_t object. Also, make sure we get the buffer
count right, not all data might have been consumed yet when
precache is called. Remove the precache/read loop in the non-sparse
case, we have already established that there is data available. If
it is insufficient, the user will simply call precache again once
it's used up, which istream_get_buffered_data forwards to a precache
call in the underlying stream.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>