Overhaul sqfs_istream_t/sqfs_ostream_t error handling
Report an error number from the implementations, change the users to
forward that error number (which also means libtar write header/link
now returns an error code) and all subsequent binaries to use
sqfs_perror() instead of relying on the function to print an error
internally.

Also, make sure to preserve errno/GetLastError() in the implementations
and print out a stringified error in sqfs_perror() if the error code
indicates an I/O error.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>