Add a hard link detecting/filtering directory iterator
The reason this is implemented separately, instead of roling it into
the recursive iterator, is so that we can do additional filtering
in between. For instance, we can rewrite the path and the hard link
path will match up, or if we remove nodes from the hierarchy, we
won't end up with a hard link pointing outside.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>