libcommon: get rid of write_data_from_file
Modify gensquashfs to pack data using an istream wrapper, similar
to tar2sqfs. To implement the no-tail-pack option, we need the file
size, so we simply open a raw handle first, and query it (using
libsquashfs API) and then create the stream wrapper. For the output
side, we create a block-processor ostream wrapper and splice it.

Since gensquashfs is the only, remaining user of the pack-a-file
functon, we can then remove it from libcommon.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>