Remove tar2sqfs test script
The whole idea of re-generating archives and matching them against
checksums has a big issue in that we cannot controll the underlying
compressors. If the underlying compressor changes its behavior
between versions, the checksums will be broken. For this exact reason,
the zstd and LZ4 checks that used to be present, have already been
removed. On Fedora, the script now breaks because of the zlib to
zstd-ng transition. If we fix the checksums to work on Fedora, it WILL
now break on other distributions. Ultimately, it is better to remove
the script alltogether.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>